Move Settings to bottom of applist and add control panel to Search
Now Search will also have Settings/Start/Pause/Shutdown qube
options; in order to avoid weird problems Settings were moved to the
bottom of the list.
